requirements.txt
psycopg2-binary>=2.9
patroni[etcd3]>=3.0